refactor(arel,activemodel,activerecord): converge mysql/sqlite visitor assertions, split TS-only type extras, drop check-constraint name quoting
Mirrors visitors/mysql_test.rb and visitors/sqlite_test.rb line by line,
moves the remaining TS-only ActiveModel type extras to .trails.test.ts
siblings, and interpolates the check-constraint name bare per
schema_creation.rb:125-127.
