+++ packages/activerecord/src/connection-adapters/abstract/schema-creation.ts
@@ -391,7 +391,7 @@ export class SchemaCreation {
   }
 
   protected visitCheckConstraintDefinition(o: CheckConstraintDefinition): string {
-    return `CONSTRAINT ${this.adapter.quoteColumnName(o.name)} CHECK (${o.expression})`;
+    return `CONSTRAINT ${o.name} CHECK (${o.expression})`;
   }
 
   async addColumnOptions(sql: string, options: ColumnOptions): Promise<string> {
