+++ packages/arel/src/visitors/sqlite.test.ts
@@ -1,121 +1,76 @@
-import { describe, it, expect } from "vitest";
+import { describe, it, expect, beforeEach } from "vitest";
 import { testConnection } from "../test-helpers/connection.js";
-import { Table, star, Nodes, Visitors } from "../index.js";
+import { mustBeLike } from "../test-helpers/must-be-like.js";
+import { buildQuoted } from "../nodes/casted.js";
+import { Table, sql, Nodes, Visitors, Collectors } from "../index.js";
 
 describe("SqliteTest", () => {
-  const users = new Table("users");
-  const posts = new Table("posts");
-  describe("Nodes::IsDistinctFrom", () => {
-    it("should handle nil", () => {
-      const node = users.get("name").isDistinctFrom(null);
-      const sql = new Visitors.SQLite(testConnection).compile(node);
-      expect(sql).toBe('"users"."name" IS NOT NULL');
-    });
+  let visitor: Visitors.SQLite;
+  beforeEach(() => {
+    visitor = new Visitors.SQLite(testConnection);
   });
 
+  function compile(node: Nodes.Node): string {
+    return visitor.accept(node, new Collectors.SQLString()).value;
+  }
+
   it("defaults limit to -1", () => {
-    const mgr = users.project(star).skip(5);
-    const sql = new Visitors.SQLite(testConnection).compile(mgr.ast);
-    expect(sql).toContain("LIMIT -1");
-    expect(sql).toContain("OFFSET 5");
+    const stmt = new Nodes.SelectStatement();
+    stmt.offset = new Nodes.Offset(1);
+    const sql = visitor.accept(stmt, new Collectors.SQLString()).value;
+    expect(mustBeLike(sql)).toBe(mustBeLike("SELECT LIMIT -1 OFFSET 1"));
   });
 
   it("does not support locking", () => {
-    const mgr = users.project(star).lock();
-    const sql = new Visitors.SQLite(testConnection).compile(mgr.ast);
-    expect(sql).not.toContain("FOR UPDATE");
+    const node = new Nodes.Lock(sql("FOR UPDATE"));
+    expect(visitor.accept(node, new Collectors.SQLString()).value).toBe("");
   });
 
   it("does not support boolean", () => {
-    const visitor = new Visitors.SQLite(testConnection);
-    expect(visitor.compile(new Nodes.True())).toBe("1");
-    expect(visitor.compile(new Nodes.False())).toBe("0");
-    // `eq` wraps the raw `true` via quotedNode into a Casted node, which is
-    // how a boolean reaches the visitor; a raw `true` placed straight into
-    // Equality raises UnsupportedVisitError, as it does in Rails.
-    expect(visitor.compile(users.get("active").eq(true))).toBe('"users"."active" = 1');
+    let node: Nodes.Node = new Nodes.True();
+    expect(visitor.accept(node, new Collectors.SQLString()).value).toBe("1");
+    node = new Nodes.False();
+    expect(visitor.accept(node, new Collectors.SQLString()).value).toBe("0");
   });
 
   describe("Nodes::IsNotDistinctFrom", () => {
+    it("should construct a valid generic SQL statement", () => {
+      const test = new Table("users").get("name").isNotDistinctFrom("Aaron Patterson");
+      expect(mustBeLike(compile(test))).toBe(mustBeLike(`"users"."name" IS 'Aaron Patterson'`));
+    });
+
     it("should handle column names on both sides", () => {
-      const node = users.get("id").isNotDistinctFrom(posts.get("user_id"));
-      const sql = new Visitors.SQLite(testConnection).compile(node);
-      expect(sql).toBe('"users"."id" IS "posts"."user_id"');
+      const test = new Table("users")
+        .get("first_name")
+        .isNotDistinctFrom(new Table("users").get("last_name"));
+      expect(mustBeLike(compile(test))).toBe(
+        mustBeLike(`"users"."first_name" IS "users"."last_name"`),
+      );
     });
 
     it("should handle nil", () => {
-      const node = users.get("name").isNotDistinctFrom(null);
-      const sql = new Visitors.SQLite(testConnection).compile(node);
-      expect(sql).toBe('"users"."name" IS NULL');
-    });
-
-    it("should construct a valid generic SQL statement", () => {
-      const node = users.get("name").isNotDistinctFrom(new Nodes.Quoted(1));
-      const sql = new Visitors.SQLite(testConnection).compile(node);
-      expect(sql).toBe('"users"."name" IS 1');
+      const table = new Table("users");
+      const val = buildQuoted(null, table.get("active"));
+      const compiled = compile(new Nodes.IsNotDistinctFrom(table.get("name"), val));
+      expect(mustBeLike(compiled)).toBe(mustBeLike(`"users"."name" IS NULL`));
     });
   });
 
   describe("Nodes::IsDistinctFrom", () => {
     it("should handle column names on both sides", () => {
-      const node = users.get("id").isDistinctFrom(posts.get("user_id"));
-      const sql = new Visitors.SQLite(testConnection).compile(node);
-      expect(sql).toContain('"users"."id"');
-      expect(sql).toContain('"posts"."user_id"');
-    });
-  });
-
-  describe("set operations", () => {
-    // SQLite rejects parens around SELECT operands of UNION/INTERSECT/EXCEPT.
-    // Mirrors `sqlite.rb#infix_value_with_paren` — strip Grouping wrappers.
-    const q1 = () => users.project(star);
-    const q2 = () => users.project(star);
-
-    it("UNION strips Grouping wrapped operands", () => {
-      const node = new Nodes.Union(new Nodes.Grouping(q1().ast), new Nodes.Grouping(q2().ast));
-      const sql = new Visitors.SQLite(testConnection).compile(node);
-      expect(sql).not.toContain("((");
-      expect(sql).toContain("UNION");
-      expect(sql).toBe('( SELECT * FROM "users" UNION SELECT * FROM "users" )');
-    });
-
-    it("UNION ALL strips Grouping wrapped operands", () => {
-      const node = new Nodes.UnionAll(new Nodes.Grouping(q1().ast), new Nodes.Grouping(q2().ast));
-      const sql = new Visitors.SQLite(testConnection).compile(node);
-      expect(sql).toBe('( SELECT * FROM "users" UNION ALL SELECT * FROM "users" )');
-    });
-
-    it("INTERSECT strips Grouping wrapped operands", () => {
-      const node = new Nodes.Intersect(new Nodes.Grouping(q1().ast), new Nodes.Grouping(q2().ast));
-      const sql = new Visitors.SQLite(testConnection).compile(node);
-      expect(sql).toBe('( (SELECT * FROM "users") INTERSECT (SELECT * FROM "users") )');
-    });
-
-    it("EXCEPT strips Grouping wrapped operands", () => {
-      const node = new Nodes.Except(new Nodes.Grouping(q1().ast), new Nodes.Grouping(q2().ast));
-      const sql = new Visitors.SQLite(testConnection).compile(node);
-      expect(sql).toBe('( (SELECT * FROM "users") EXCEPT (SELECT * FROM "users") )');
-    });
-
-    it("UNION without Grouping operands renders bare SELECTs", () => {
-      const node = new Nodes.Union(q1().ast, q2().ast);
-      const sql = new Visitors.SQLite(testConnection).compile(node);
-      expect(sql).toBe('( SELECT * FROM "users" UNION SELECT * FROM "users" )');
-    });
-
-    it("nested unions are flattened (Rails-style)", () => {
-      const q3 = users.project(star);
-      const node = new Nodes.Union(q1().ast, new Nodes.Union(q2().ast, q3.ast));
-      const sql = new Visitors.SQLite(testConnection).compile(node);
-      expect(sql).toBe(
-        '( SELECT * FROM "users" UNION SELECT * FROM "users" UNION SELECT * FROM "users" )',
+      const test = new Table("users")
+        .get("first_name")
+        .isDistinctFrom(new Table("users").get("last_name"));
+      expect(mustBeLike(compile(test))).toBe(
+        mustBeLike(`"users"."first_name" IS NOT "users"."last_name"`),
       );
     });
 
-    it("union via SelectManager omits inner parens", () => {
-      const sql = new Visitors.SQLite(testConnection).compile(q1().union(q2()));
-      expect(sql).not.toContain("((");
-      expect(sql).toContain("UNION");
+    it("should handle nil", () => {
+      const table = new Table("users");
+      const val = buildQuoted(null, table.get("active"));
+      const compiled = compile(new Nodes.IsDistinctFrom(table.get("name"), val));
+      expect(mustBeLike(compiled)).toBe(mustBeLike(`"users"."name" IS NOT NULL`));
     });
   });
 });
